feat: add code to save PDF data to CSV file
The code changes in this commit add functionality to save PDF data to a CSV file. The `main.py` file in the `data_collector` directory has been modified to include the following changes:
- A print statement has been added to display the name of each PDF file being processed.
- The `add_rows` method of the `csvConstructor` object is called with the text extracted from the PDF file.
- The `save_csv` method of the `csvConstructor` object is called to save the CSV file.

This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature) and providing a concise and descriptive message about the purpose of the changes.